requirements.txt
flask>=2.0.0
flask-socketio>=5.0.0
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.25.0
langchain>=0.1.0
langgraph>=0.0.17
mistralai>=0.0.7
boto3>=1.26.0
werkzeug>=2.0.0
eventlet==0.39.1
python-engineio==4.12.0
python-socketio==5.13.0
bidict==0.23.1
cartesia>=0.2.0